Fix clustering to run on full universe before pair search
Previously get_clusters() was called after ticker_limit sliced the
universe to a random 100-ticker sample, causing HDBSCAN to re-cluster
a different random set every day. Clusters now run on all tickers
before any limit is applied so the structure reflects the full universe
and is cached correctly across days (recompute_days=None).

ticker_limit is now only applied in the non-clustered fallback path.

Also fix TickerClusterer dtype: cast prices to float64 before np.log
to handle Python float values returned from PostgreSQL.

Made-with: Cursor

+++ TickerClusterer.py
@@ -129,8 +129,10 @@ class TickerClusterer:
             return [list(tickers)]
 
         # Require ≥50% coverage; forward-fill gaps then drop any column still NaN.
+        # Cast to float64 explicitly — DB values come back as Python floats which
+        # numpy ufuncs (np.log) cannot operate on via the array protocol.
         min_obs = max(2, int(prices.shape[0] * 0.5))
-        prices = prices.dropna(axis=1, thresh=min_obs).ffill()
+        prices = prices.dropna(axis=1, thresh=min_obs).ffill().astype(float)
         log_returns = np.log(prices).diff().dropna()
         log_returns = log_returns.dropna(axis=1)
